Hoist error definition up to engine
For generality. This allows SimEngine to use it, and also allows
processing chains as appropriate.
