+++ src/engine/mod.rs
@@ -5,9 +5,10 @@
 pub use self::{
     devlinks::filesystem_mount_path,
     engine::{BlockDev, Engine, Filesystem, Pool},
+    errors::{Error, ErrorKind},
     event::{get_engine_listener_list_mut, EngineEvent, EngineListener},
     sim_engine::SimEngine,
-    strat_engine::{Error, ErrorKind, StratEngine},
+    strat_engine::StratEngine,
     types::{
         BlockDevState, BlockDevTier, DevUuid, FilesystemUuid, MaybeDbusPath, Name, PoolUuid,
         Redundancy, RenameAction,
@@ -20,6 +21,7 @@ mod macros;
 mod devlinks;
 #[allow(clippy::module_inception)]
 mod engine;
+mod errors;
 mod event;
 mod sim_engine;
 mod strat_engine;
